chore(project.inlang): setup i18n configuration
These files were auto-generated when I first ran `pnpm dev`. I'm not
sure if they should be ignored or not but they seem safe to commit. I
tried to find further documentation on them but could not. I suspect
they are related to the `paraglide` setup that `npx sv create` did.