fix: get correct public key for JWT verification
Verify the JWT using the getKey callback with the `jsonwebtoken`
library. This can be used for any JWT, I believe. Use `node-jwks-rsa` as
a way to fetch the public keys used in verification.

I could probably also just use the Node.js `google-auth-library` but
this is kind of fun and just as easy.

Ref: https://github.com/googleapis/google-auth-library-nodejs
Ref: https://developers.google.com/identity/gsi/web/guides/verify-google-id-token#node.js
Ref: https://www.npmjs.com/package/jsonwebtoken#jwtverifytoken-secretorpublickey-options-callback
Ref: https://github.com/auth0/node-jwks-rsa

Fixes: ae739c7 ("fix(server/google): get public keys every time")