chore: use `just` as primary command runner
Remove most of the special scripts from `package.json` in favor of a
dedicated `Justfile` that contains those recipes and descriptions of
them in comments. This is a much better developer experience than simple
`pnpm run` scripts. `just` has many more features. Even just the listing
with descriptions and groups makes this change worth it.